--- components/notification/Header.module.scss
@@ -1,7 +1,0 @@
-@import "~hds-design-tokens/lib/all.scss";
-
-.notificationHeader {
-  .navigationItem:after {
-    background-color: $color-bus !important;
-  }
-}
