+++ components/notification/NotificationHeader.module.scss
@@ -1,0 +1,7 @@
+@import "~hds-design-tokens/lib/all.scss";
+
+
+.navigationItem:after {
+  background-color: $color-bus !important;
+}
+
